requirements.txt
sqlglot>=25.0
aiohttp>=3.9
pyyaml>=6.0
scikit-learn>=1.4
numpy>=1.26
